requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
boto3==1.26.93
botocore==1.29.93
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.1
defusedxml==0.7.1
Django==4.2
django-allauth==0.54.0
django-environ==0.10.0
frozenlist==1.3.3
geoip2==4.6.0
idna==3.4
jmespath==1.0.1
maxminddb==2.2.0
multidict==6.0.4
mysqlclient==2.1.1
oauthlib==3.2.2
pip==23.0.1
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
requests==2.28.2
requests-oauthlib==1.3.1
s3transfer==0.6.0
setuptools==65.6.3
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
urllib3==1.26.15
wheel==0.38.4
wincertstore==0.2
yarl==1.8.2
